don't show bsc banner on the home page
